injsite/requirements.txt
appdirs==1.4.2
dj-database-url==0.4.2
Django==1.10.6
django-registration==2.2
gunicorn==19.7.0
packaging==16.8
psycopg2==2.7
pyparsing==2.1.10
six==1.10.0
whitenoise==3.3.0
